# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZipLink API Documentation",
        "description": "Высокопроизводительный сервис сокращения ссылок и сбора аналитики",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Основной API сервер"
        }
    ],
    "paths": {
...
        "/tokens": {
            "get": {
                "tags": [
                    "API Tokens"
                ],
                "summary": "Получение списка активных API-токенов",
                "operationId": "eb66ab14ef6279023894ef3391698c5c",
                "responses": {
                    "200": {
                        "description": "Список токенов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Bot Token"
                                                    },
                                                    "last_used_at": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2026-09-19T22:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "API Tokens"
                ],
                "summary": "Выпуск нового персонального API-токена",
                "operationId": "699699145f4e71e9ed09da22a7dcac0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "CI/CD Token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Токен создан (показывается только 1 раз)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token created successfully."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef123456..."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "API Tokens"
                ],
                "summary": "Отозвать все токены текущего пользователя",
                "operationId": "2b2d9f3b72042461228e65a44fe70611",
                "responses": {
                    "200": {
                        "description": "Все токены успешно отозваны"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/tokens/{tokenId}": {
            "delete": {
                "tags": [
                    "API Tokens"
                ],
                "summary": "Отозвать конкретный токен по ID",
                "operationId": "b71924377d4ef786c3a929d791d60c23",
                "parameters": [
                    {
                        "name": "tokenId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Токен отозван"
                    },
                    "404": {
                        "description": "Токен не найден"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Введите токен в формате: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}